pyproject.toml
[tool.poetry]
name = "myproject"
version = "0.1.0"
description = ""
authors = ["Kawakubo322 <tetsu.kawakubo0322@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
yapf = "^0.31.0"
isort = "^5.8.0"
pyocr = "^0.8"
Pillow = "^8.2.0"
click = "^7.1.2"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
